+++ zhmccli/_cmd_partition.py
@@ -28,11 +28,11 @@ import zhmcclient
 from .zhmccli import cli, CONSOLE_LOGGER_NAME
 from ._helper import print_properties, print_resources, abort_if_false, \
     options_to_properties, original_options, COMMAND_OPTIONS_METAVAR, \
-    part_console, raise_click_exception, add_options, LIST_OPTIONS
+    part_console, raise_click_exception, storage_management_feature, \
+    add_options, LIST_OPTIONS
 from ._cmd_cpc import find_cpc
 from ._cmd_storagegroup import find_storagegroup
 from ._cmd_metrics import get_metric_values
-
 
 # Defaults for partition creation
 DEFAULT_IFL_PROCESSORS = 1
@@ -142,6 +142,58 @@ def partition_stop(cmd_ctx, cpc, partition):
     'zhmc' command name.
     """
     cmd_ctx.execute_cmd(lambda: cmd_partition_stop(cmd_ctx, cpc, partition))
+
+
+@partition_group.command('dump', options_metavar=COMMAND_OPTIONS_METAVAR)
+@click.argument('CPC', type=str, metavar='CPC')
+@click.argument('PARTITION', type=str, metavar='PARTITION')
+@click.option('--volume', type=str, required=True,
+              help='The storage volume that contains the dump program. '
+              'For CPCs with the storage management feature (z14 and later): '
+              'A string of the form "SG.SV" where SG is the name '
+              'of the storage group resource attached to the partition and SV '
+              'is the name of the storage volume in that storage group, or of '
+              'the form "UUID" where UUID is the UUID of the storage volume '
+              'on the storage array (also shown in the HMC GUI). The storage '
+              'volume may be of type FCP or FICON. '
+              'For CPCs without the storage management feature (z13 and '
+              'earlier): A string of the form "HBA.WWPN.LUN", where HBA is the '
+              'name of the HBA resource in the partition and WWPN and LUN '
+              'identify the storage array and storage volume thereon. The '
+              'storage volume must be of type FCP.')
+@click.option('--lba', type=str, required=False, default='0',
+              help='The logical block number of the anchor point for locating '
+              'the dump program on the storage volume. Default: 0')
+@click.option('--configuration', type=int, required=False, default=0,
+              help='A selector that is passed to the dump program for '
+              'selecting the boot configuration to use. 0 means to use the '
+              'default boot configuration. Default: 0')
+@click.option('--parameters', type=str, required=False, default='',
+              help='A parameter string that is passed to the dump program as '
+              'boot parameters. Default: empty string')
+@click.option('--timeout', type=int, required=False, default=60,
+              help='The time in seconds that is waited before the load of the '
+              'dump program from a FICON storage volume is aborted. Only used '
+              'for CPCs with the storage management feature (z14 and later) '
+              'when loading from a FICON storage volume, and ignored '
+              'otherwise. Default: 60')
+@click.pass_obj
+def partition_dump(cmd_ctx, cpc, partition, **options):
+    """
+    Load and execute a standalone dump program from a volume.
+
+    This command loads a standalone dump program into a partition and begins
+    its execution. It does so in a special way that the existing contents of
+    the partition's memory are not overwritten so that the dump program can
+    dump those contents. The partition must be in one of the states "active",
+    "degraded", "paused", or "terminated".
+
+    In addition to the command-specific options shown in this help text, the
+    general options (see 'zhmc --help') can also be specified right after the
+    'zhmc' command name.
+    """
+    cmd_ctx.execute_cmd(lambda: cmd_partition_dump(
+        cmd_ctx, cpc, partition, **options))
 
 
 @partition_group.command('create', options_metavar=COMMAND_OPTIONS_METAVAR)
@@ -984,6 +1036,133 @@ def cmd_partition_console(cmd_ctx, cpc_name, partition_name, options):
     except zhmcclient.Error as exc:
         raise click.ClickException(
             "{exc}: {msg}".format(exc=exc.__class__.__name__, msg=exc))
+
+
+def cmd_partition_dump(cmd_ctx, cpc_name, partition_name, **options):
+    # pylint: disable=missing-function-docstring
+
+    client = zhmcclient.Client(cmd_ctx.session)
+    partition = find_partition(cmd_ctx, client, cpc_name, partition_name)
+
+    name_map = {
+        # These parameters are the same for CPCs with and without the storage
+        # management feature. Feature-specific parameters are added further
+        # down:
+        'configuration': 'dump-configuration-selector',
+        'parameters': 'dump-os-specific-parameters',
+        'lba': 'dump-record-lba',
+
+        # Options that are handled in this function:
+        'volume': None,
+        'timeout': None,
+    }
+    options = original_options(options)
+    properties = options_to_properties(options, name_map)
+    volume_opt = options['volume']  # It is required
+    timeout = options['timeout']  # It is optional but has a default
+
+    if storage_management_feature(partition):
+
+        # Check and parse the --volume option into SG.SV or UUID
+        volume_parts = volume_opt.split('.')
+        if len(volume_parts) == 1:  # format: UUID
+            uuid = volume_parts[0]
+            sg_list = partition.list_attached_storage_groups()
+            storage_volume = None
+            for sg in sg_list:
+                try:
+                    storage_volume = sg.storage_volumes.find(uuid=uuid)
+                    break
+                except zhmcclient.NotFound:
+                    continue
+            if not storage_volume:
+                raise_click_exception(
+                    "Storage volume with UUID '{uuid}' specified in the "
+                    "--volume option was not found in any storage group "
+                    "attached to partition '{part}' of CPC '{cpc}'.".
+                    format(uuid=uuid, part=partition_name, cpc=cpc_name),
+                    cmd_ctx.error_format)
+        elif len(volume_parts) == 2:  # format: SG.SV
+            sg_name, sv_name = volume_parts
+            sg_list = partition.list_attached_storage_groups()
+            sg = None
+            for _sg in sg_list:
+                if _sg.name == sg_name:
+                    sg = _sg
+                    break
+            if not sg:
+                raise_click_exception(
+                    "Storage group '{sg}' specified in the --volume option was "
+                    "not found in the storage groups attached to partition "
+                    "'{part}' of CPC '{cpc}'.".
+                    format(sg=sg_name, part=partition_name, cpc=cpc_name),
+                    cmd_ctx.error_format)
+            try:
+                storage_volume = sg.storage_volumes.find(name=sv_name)
+            except zhmcclient.NotFound:
+                raise_click_exception(
+                    "Storage volume '{sv}' specified in the --volume option "
+                    "was not found in the volumes of storage group '{sg}' "
+                    "attached to partition '{part}' of CPC '{cpc}'.".
+                    format(sv=sv_name, sg=sg_name, part=partition_name,
+                           cpc=cpc_name),
+                    cmd_ctx.error_format)
+        else:
+            raise_click_exception(
+                "Invalid format for --volume option: '{v}'. CPC '{cpc}' has "
+                "the storage management feature which requires the formats "
+                "'SG.SV' or 'UUID'.".
+                format(v=volume_opt, cpc=cpc_name),
+                cmd_ctx.error_format)
+
+        # Set the remaining feature-specific parameters
+        properties['storage-volume-uri'] = storage_volume.uri
+        properties['timeout'] = timeout
+
+        op_properties = {
+            'dump-program-info': properties,
+            'dump-program-type': 'storage',
+        }
+        try:
+            partition.start_dump_program(
+                parameters=op_properties, wait_for_completion=True)
+        except zhmcclient.Error as exc:
+            raise_click_exception(exc, cmd_ctx.error_format)
+
+    else:  # No storage management feature
+
+        # Check and parse the --volume option into HBA.WWPN.LUN
+        volume_parts = volume_opt.split('.')
+        if len(volume_parts) != 3:
+            raise_click_exception(
+                "Invalid format for --volume option: '{v}'. CPC '{cpc}' does "
+                "not have the storage management feature which requires the "
+                "format 'HBA.WWPN.LUN'.".
+                format(v=volume_opt, cpc=cpc_name),
+                cmd_ctx.error_format)
+        hba_name, wwpn, lun = volume_parts
+        try:
+            hba = partition.hbas.find(name=hba_name)
+        except zhmcclient.NotFound:
+            raise_click_exception(
+                "HBA '{hba}' specified in the --volume option was not found "
+                "in partition '{part}' of CPC '{cpc}'.".
+                format(hba=hba_name, part=partition_name, cpc=cpc_name),
+                cmd_ctx.error_format)
+
+        # Set the remaining feature-specific parameters
+        properties['dump-load-hba-uri'] = hba.uri
+        properties['dump-world-wide-port-name'] = wwpn
+        properties['dump-logical-unit-number'] = lun
+
+        try:
+            partition.dump_partition(
+                parameters=properties, wait_for_completion=True)
+        except zhmcclient.Error as exc:
+            raise_click_exception(exc, cmd_ctx.error_format)
+
+    cmd_ctx.spinner.stop()
+    click.echo('Dump of Partition %s is complete.' % partition_name)
 
 
 def cmd_partition_mount_iso(cmd_ctx, cpc_name, partition_name, options):